requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-bootstrap5==23.3
djangorestframework==3.14.0
firebase-admin==6.3.0
gcloud==0.18.3
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
jws==0.1.3
msgpack==1.0.7
mysqlclient==2.2.0
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==2.0.1
pytz==2023.3.post1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
